pillow/requirements.windows.txt
Pillow>=8.1.2
win32-setctime>=1.0.3